[dev.cc] cmd/asm: rewrite to work with new obj API
Considerable rewriting of the parser and assembler (code generator)
but it's simpler and shorter now. The internal Addr type is gone; so
is the package that held it. Parsing of operands goes directly into
obj.Addrs now.

There is a horrible hack regarding register pairs. It uses the Class
field to store the second register since it needs _some_ place to
put it but none is provided in the API. An alternative would be nice
but this works for now.

Once again creates identical .6 and .8 files as the old assembler.

Reviewed-on: https://go-review.googlesource.com/4062

+++ src/cmd/asm/internal/lex/lex.go
@@ -61,8 +61,9 @@ var (
 // HistLine reports the cumulative source line number of the token,
 // for use in the Prog structure for the linker. (It's always handling the
 // instruction from the current lex line.)
-func HistLine() int {
-	return histLine
+// It returns int32 because that's what type ../asm prefers.
+func HistLine() int32 {
+	return int32(histLine)
 }
 
 // NewLexer returns a lexer for the named file and the given link context.
